Refactor path resolution utilities
