+++ jac-splice-orc/jac_splice_orc/plugin/splice_plugin.py
@@ -16,6 +16,7 @@ from jac_splice_orc.managers.proxy_manager import ModuleProxy, RemoteObjectProxy
 
 from jaclang.cli.cmdreg import cmd_registry
 from jaclang.runtimelib.machine import JacMachine, JacMachineInterface, JacProgram
+from jaclang.utils import infer_language
 
 
 from kubernetes import client, config
@@ -591,7 +592,7 @@ class SpliceOrcPlugin:
             PythonImporter,
         )
 
-        lng = JacMachine.infer_language(target, base_path)
+        lng = infer_language(target, base_path)
         module_config_path = os.getenv("MODULE_CONFIG_PATH", "/cfg/module_config.json")
         try:
             logging.debug(f"Loading from {module_config_path} for module_config...")
