+++ jac/jaclang/utils/__init__.py
@@ -1,1 +1,3 @@
 """Jaseci utility functions and libraries."""
+
+from .module_resolver import resolve_module, infer_language, resolve_relative_path
